predict.py

+++ account-service/src/main/java/org/example/dto/AccountDto.java
@@ -3,6 +3,9 @@ package org.example.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -12,6 +15,6 @@ public class AccountDto {
     private String userId;
     private String accountType;
     private String accountNumber;
-    private Double balance;
+    private BigDecimal balance;
     private String currency;
 }
